Fixed base XP gain calculation at mob kill. Patch provided by BigBear.
git-svn-id: http://mangos.svn.sourceforge.net/svnroot/mangos@2151 26674162-90ff-0310-a85b-b64107c8dee9

+++ trunk/src/game/Formulas.h
@@ -182,7 +182,7 @@ namespace MaNGOS
                 if( mob_level > gray_level )
                 {
                     uint32 ZD = GetZeroDifference(pl_level);
-                    return ( (pl_level*5 + 45) * (1 - (pl_level - mob_level)/ZD) );
+                    return uint32( (pl_level*5 + 45) * (1 - float(pl_level - mob_level)/ZD) );
                 }
                 return 0;
             }
